% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.extratGO}
\alias{.extratGO}
\title{extract GO information from NCBI and filter by taxid}
\usage{
.extratGO(taxid = NULL, species = NULL)
}
\arguments{
\item{taxid}{taxonomy id for the species}

\item{species}{species name(common name,kegg.species.code or scientifc name)}
}
\value{
dataframe with gene2go information
}
\description{
extract GO information from NCBI and filter by taxid
}
\author{
Kai Guo
}
