% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get.species.info}
\alias{.get.species.info}
\title{modified from pathview kegg.species.code}
\usage{
.get.species.info(species = "hsa", na.rm = FALSE)
}
\arguments{
\item{species}{species name(common name,kegg.species.code or scientifc name)}

\item{na.rm}{TRUE/FALSE}
}
\description{
modified from pathview kegg.species.code
}
