% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.getdbname}
\alias{.getdbname}
\title{get database name by using species name}
\usage{
.getdbname(species = species)
}
\arguments{
\item{species}{species name}
}
\description{
get database name by using species name
}
\author{
Kai Guo
}
