% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.getmartdb}
\alias{.getmartdb}
\title{get species information in Ensembl}
\usage{
.getmartdb(species, mart)
}
\arguments{
\item{species}{species}

\item{mart}{biomaRt mart}
}
\description{
get species information in Ensembl
}
\author{
Kai Guo
}
