% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.show.path}
\alias{.show.path}
\title{show the package path}
\usage{
.show.path(package)
}
\arguments{
\item{package}{the full path of the package}
}
\value{
whole path for the package
}
\description{
show the package path
}
\author{
Kai Guo
}
