% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.show.tables}
\alias{.show.tables}
\title{show the package content}
\usage{
.show.tables(package)
}
\arguments{
\item{package}{the full path of the package}
}
\value{
vector
}
\description{
show the package content
}
\author{
Kai Guo
}
