% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnsembl.R
\name{fromEnsembl}
\alias{fromEnsembl}
\title{build annotation from ensembl}
\usage{
fromEnsembl(
  species = "Arabidopsis t",
  host = "uswest.ensembl.org",
  anntype = NULL,
  buildall = TRUE,
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  plant = FALSE,
  install = TRUE,
  outputDir = NULL
)
}
\arguments{
\item{species}{the sepcies you want to search, you can use listSpecies to get the species name}

\item{host}{the ensemble API host,for plant you can use plants.ensembl.org and for human and other species you can use uswest.ensembl.org}

\item{anntype}{the type of function annotation(GO,KEGG,PFAM,InterPro) you want get from ensemble}

\item{buildall}{include all prossbile annoation type listed in Ensembl}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus name}

\item{version}{version number(xx.xx.xx)}

\item{plant}{plant or animal species (TRUE/FALSE)}

\item{install}{install the package or not(default: TRUE)}
}
\description{
build annotation from ensembl
}
\examples{
\dontrun{
fromEnsembl(species="Human")
}
}
\author{
Kai Guo
}
