% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKEGG.R
\name{fromKEGG}
\alias{fromKEGG}
\title{make annotation database from KEGG and GO from NCBI}
\usage{
fromKEGG(
  species = "ath",
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  install = TRUE,
  outputDir = NULL,
  rebuild = FALSE
)
}
\arguments{
\item{species}{species name(common name,kegg.species.code or scientifc name)}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer for the annotation package}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{version}{version for the annotation package}

\item{install}{install the package or not(default: TRUE)}

\item{outputDir}{temporary output path}

\item{rebuild}{rebuild the package or not(default: FALSE)}
}
\value{
annotation package
}
\description{
make annotation database from KEGG and GO from NCBI
}
\examples{
donttest{
fromKEGG(species="ath")
}
}
\author{
Kai Guo
}
