% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNCBI.R
\name{fromNCBI}
\alias{fromNCBI}
\title{build annotation database from NCBI}
\usage{
fromNCBI(
  species = "ath",
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  install = TRUE,
  outputDir = NULL,
  rebuild = FALSE
)
}
\arguments{
\item{species}{species name}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer for the annotation package}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{version}{version for the annotation package}

\item{install}{install the package or not(default: TRUE)}

\item{outputDir}{temporary output path}

\item{rebuild}{rebuild the package or not(default: FALSE)}
}
\value{
annotation package
}
\description{
build annotation database from NCBI
}
\examples{
\donttest{
## build annoataion package for Ecoli
fromNCBI(species = "eco", install = FALSE)
}
}
\author{
Kai Guo
}
