% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOwn.R
\name{fromOwn}
\alias{fromOwn}
\title{build Own annotation database with user defined annotation file}
\usage{
fromOwn(
  geneinfo = geneinfo,
  gene2go = NULL,
  gene2path = NULL,
  version = NULL,
  maintainer = NULL,
  author = NULL,
  outputDir = NULL,
  tax_id = NULL,
  genus = NULL,
  species = NULL,
  install = TRUE
)
}
\arguments{
\item{geneinfo}{gene information table with two columns
as default("GID","DESCRIPTION")}

\item{gene2go}{Gene Onotoly information for  genes}

\item{gene2path}{KEGG Pathway or KO information for genes}

\item{version}{version for the annotation package}

\item{maintainer}{maintainer for the annotation package}

\item{author}{author for the annotation package}

\item{outputDir}{temporary output path}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{species}{species name(common name,kegg.species.code or scientifc name)}

\item{install}{install the package or not(default: TRUE)}
}
\value{
annotation package
}
\description{
build Own annotation database with user defined annotation file
}
\examples{
## build your own annotation for Arabidopsis thaliana
data(ath)
fromOwn(geneinfo = ath, install = FALSE)
}
\author{
Kai Guo
}
