% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getTable}
\alias{getTable}
\title{get annotataion table from temporary package}
\usage{
getTable(path, table = "go_all")
}
\arguments{
\item{path}{full path for the temporary package}

\item{table}{a character  indicate the table you want extract}
}
\value{
data.frame
}
\description{
get annotataion table from temporary package
}
\author{
Kai Guo
}
