% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{korg}
\alias{korg}
\title{korg}
\format{A matrix with five columns:
\describe{
\item{ktax.id}{KEGG taxonomy ID}
\item{tax.id}{NCBI taxonomy ID}
\item{kegg.code}{KEGG species code}
\item{scientific.name}{Scientific name of species}
\item{common.name}{common name of species}
}}
\usage{
korg
}
\description{
korg include species information from KEGG database.
             korg data was modified from
             (https://pathview.uncc.edu/data/korg.tsv)
}
\examples{
korg

}
\keyword{datasets}
