% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnsembl.R
\name{listSpecies}
\alias{listSpecies}
\title{list species available in Ensembl}
\usage{
listSpecies(host = "uswest.ensembl.org", plant = FALSE)
}
\arguments{
\item{host}{Ensembl host site}
}
\description{
list species available in Ensembl
}
\author{
Kai Guo
}
