% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDB.R
\name{mergeDB}
\alias{mergeDB}
\title{merge two orgDB with keys}
\usage{
mergeDB(
  dbleft,
  dbright,
  keyleft = "GID",
  keyright = "GID",
  keytype = NULL,
  keep = NULL,
  species = NULL,
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  pkgname = NULL,
  install = TRUE,
  outputDir = NULL,
  rebuild = FALSE
)
}
\arguments{
\item{dbleft}{a charater indicate the left orgDB}

\item{dbright}{a character indicate the right orgDB}

\item{keyleft}{the keytype use for merging in left orgDB}

\item{keyright}{the keytype use for merging in the right orgDB}

\item{keytype}{the keytypes to be included in the merged orgDB ("GID","GENENAME")}

\item{keep}{the name of keytype you used if keyleft and keyright were not same}

\item{species}{the species name}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer for the annotation package}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus name for the annotation package}

\item{version}{version number for the annotation package}

\item{pkgname}{package name you want to choose}

\item{install}{install the package or not(default: TRUE)}

\item{outputDir}{temporary output path}

\item{rebuild}{rebuild the package or not(default: FALSE)}
}
\description{
merge two orgDB with keys
}
\examples{
fromKEGG(species = "hsa", anntype="KEGG")
fromAnnHub(species="human")
mergeDB("org.hsa.eg.db","org.human.eg.db",species="merge")
}
\author{
Kai Guo
}
