% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath}
\alias{ath}
\title{TAIR10 geneid to ENTREZID}
\format{
A data.frame with two columns: 
\describe{
\item{GID}{The arabidopsis GENE ID}
\item{ENTREZID}{NCBI ENTREZID ID for the arabidopsis}
}
}
\usage{
ath
}
\description{
The 'ath' dataset include the annotation information collected
form the TAIR10 database(htps://arabidopsis.org/download/index-auto.jsp
%3Fdir%3D%252Fdownload_files%252FGenes%252FTAIR10_genome_release).
}
\examples{
head(ath)

}
\keyword{datasets}
