% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath}
\alias{ath}
\title{TAIR10 geneid to ENTREZID}
\format{A matrix with two columns:
\describe{
\item{GID}{TAIR GENE ID}
\item{ENTREZID}{NCBI ENTREZID ID}
}}
\usage{
ath
}
\description{
korg include species information from KEGG database. korg data was modified from
             (https://pathview.uncc.edu/data/korg.tsv)
}
\examples{
\dontrun{
ath
}
}
\keyword{datasets}
