% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.extratGO}
\alias{.extratGO}
\title{extract GO information from NCBI and filter by taxid}
\usage{
.extratGO(taxid = NULL, sepecies = NULL)
}
\description{
extract GO information from NCBI and filter by taxid
}
\author{
Kai Guo
}
