% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAnntationHub.R
\name{fromAnnHub}
\alias{fromAnnHub}
\title{extract annotation database by using AnnotationHub}
\usage{
fromAnnHub(
  species,
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  install = TRUE,
  outputDir = NULL,
  rebuild = FALSE
)
}
\arguments{
\item{species}{species name(common name,kegg.species.code or scientifc name)}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer for the annotation package}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{version}{version for the annotation package}

\item{install}{install the package or not(default: TRUE)}

\item{outputDir}{temporary file path}

\item{rebuild}{rebuild the package or not(default: FALSE)}
}
\value{
annotation package
}
\description{
extract annotation database by using AnnotationHub
}
\examples{
require(bioAnno)
fromAnnhub(species="ath")
}
\author{
Kai Guo
}
