% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAnntationHub.R
\name{fromAnnHub}
\alias{fromAnnHub}
\title{extract annotation database by using AnnotationHub}
\usage{
fromAnnHub(species, useEnsembl = FALSE, useNCBI = FALSE, useUCSC = FALSE)
}
\description{
extract annotation database by using AnnotationHub
}
\author{
Kai Guo
}
