% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNCBI.R
\name{fromNCBI}
\alias{fromNCBI}
\title{make annotation database from NCBI}
\usage{
fromNCBI(
  species = "ath",
  author = NULL,
  maintainer = NULL,
  tax_id = NULL,
  genus = NULL,
  version = NULL,
  install = TRUE,
  outputDir = NULL
)
}
\arguments{
\item{species}{species name}

\item{author}{author for the annotation package}

\item{maintainer}{maintainer}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{version}{version}

\item{install}{install the package or not(default: TRUE)}

\item{plant}{plant or animal species (TRUE/FALSE)}
}
\description{
make annotation database from NCBI
}
\examples{
\dontrun{
fromKEGG(species="ath")
}
}
\author{
Kai Guo
}
