% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildOwn.R
\name{fromOwn}
\alias{fromOwn}
\title{build Own annotation database with user defined annotation file}
\usage{
fromOwn(
  geneinfo = geneinfo,
  gene2go = NULL,
  gene2path = NULL,
  version = NULL,
  maintainer = NULL,
  author = NULL,
  outputDir = NULL,
  tax_id = NULL,
  genus = NULL,
  species = NULL
)
}
\arguments{
\item{geneinfo}{gene information table with two columns as default("GID","DESCRIPTION")}

\item{gene2go}{Gene Onotoly information for  genes}

\item{gene2path}{KEGG Pathway or KO information for genes}

\item{version}{version}

\item{maintainer}{maintainer}

\item{author}{author for the annotation package}

\item{tax_id}{taxonomy id for the species}

\item{genus}{genus for the species}

\item{entrezid}{gene information table with corresponding ENTREZID id}

\item{refseq}{gene information table with corresponding REFSEQ id}

\item{install}{install the package or not(default: TRUE)}
}
\description{
build Own annotation database with user defined annotation file
}
